+++ packages/node/src/indexer/dynamic-ds.service.ts
@@ -26,15 +26,15 @@ export class DynamicDsService extends BaseDynamicDsService<SubstrateProjectDs> {
   protected async getDatasource(
     params: DatasourceParams,
   ): Promise<SubstrateProjectDs> {
-    const { name, ...template } = cloneDeep(
-      this.project.templates.find((t) => t.name === params.templateName),
+    const t = this.project.templates.find(
+      (t) => t.name === params.templateName,
     );
-
-    if (!template) {
+    if (!t) {
       throw new Error(
         `Unable to find matching template in project for name: "${params.templateName}"`,
       );
     }
+    const { name, ...template } = cloneDeep(t);
 
     const dsObj = {
       ...template,